main_svc 8

+++ stats/dto/src/main/java/ru/practicum/dto/ViewStats.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class ViewStats {
     private String app;
     private String uri;
-    private Integer hits;
+    private long hits;
 
     @Override
     public String toString() {
